further organize files into appropriate folders
