+++ src/electricity/parsers/__init__.py
@@ -1,0 +1,4 @@
+from REDD_Parser import Redd_Parser
+from Refit_Parser import Refit_Parser
+from UKDALE_Parser import UK_Dale_Parser
+from PChainger_Parser import PChainger_Parser
